+++ HarryPotterBook/HarryPotterBook/Core/Entity/Book.swift
@@ -13,6 +13,10 @@ struct Book{
     let pages: Int
     let releaseDate: String
     let dedication: String
-    let wiki: URL
+    let summary: String
+    let wiki: URL?
     let chapters: [String]
 }
+
+// Date 타입 형식 바꾸고
+// Book Entity에 이미지도 삽입되도록
