+++ HarryPotterBook/HarryPotterBook/Core/Service/DataService.swift
@@ -22,7 +22,7 @@ final class DataService {
         
         do {
             let data = try Data(contentsOf: URL(fileURLWithPath: path))
-            let bookResponse = try JSONDecoder().decode(BookResponse.self, from: data)
+            let bookResponse = try JSONDecoder().decode(BookResponseDTO.self, from: data)
             let books = bookResponse.data.map { $0.attributes.toEntity() }
             completion(.success(books))
         } catch {
